requirements.txt
dotenv==0.9.9
python-dotenv==1.0.1
python-multipart==0.0.20
dropbox==12.0.2
fastapi==0.115.11
uvicorn==0.34.0
bcrypt==4.0.0
passlib==1.7.4
python-jose==3.3.0
email_validator==2.1.1
SQLAlchemy==1.4.40
psycopg2==2.9.10
PyJWT
pydantic==2.11.2
openai==1.66.3
